[ISSUE-34] Refactoring tools maven project into three: tools-api, tools-cli & tools-mvn
