+++ tools/api/pom.xml
@@ -15,26 +15,27 @@
   ~ limitations under the License.
   -->
 <project
-        xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.wildfly-extras</groupId>
-        <artifactId>batavia-parent</artifactId>
+        <artifactId>transformer-tools-parent</artifactId>
         <version>1.0.0.Alpha1-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <name>Transformer Tools Parent</name>
-    <artifactId>transformer-tools-parent</artifactId>
-    <packaging>pom</packaging>
+    <name>Transformer Tools API</name>
+    <artifactId>transformer-tools-api</artifactId>
+    <packaging>jar</packaging>
 
-    <modules>
-        <module>api</module>
-        <module>cli</module>
-        <module>mvn</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.wildfly-extras</groupId>
+            <artifactId>transformer-api</artifactId>
+        </dependency>
+    </dependencies>
 
 </project>
